feat(checkbox)!: remove top level subcomponents
