+++ src/components/Checkbox/index.ts
@@ -1,2 +1,1 @@
 export { Checkbox as default } from './Checkbox';
-export type { CheckboxProps } from './Checkbox';
