--- src/components/CheckboxInput/index.ts
@@ -1,2 +1,0 @@
-export { CheckboxInput as default } from './CheckboxInput';
-export type { CheckboxInputProps } from './CheckboxInput';
