--- src/components/CheckboxLabel/CheckboxLabel.module.css
@@ -1,5 +1,0 @@
-.checkbox__label--md {
-  /* Center the first line of the medium label with the checkbox. */
-  position: relative;
-  top: 0.0625rem;
-}
