--- src/components/CheckboxLabel/index.ts
@@ -1,2 +1,0 @@
-export { CheckboxLabel as default } from './CheckboxLabel';
-export type { CheckboxLabelProps } from './CheckboxLabel';
